decompress: update various jpeg 2000 standard and library limits

+++ src/lib/jp2/t1/Codeblock.h
@@ -261,7 +261,7 @@ struct DecompressCodeblock : public Codeblock
 
   private:
 	Segment* segs; /* information on segments */
-	uint8_t numSegments; /* number of segment in block*/
+	uint16_t numSegments; /* number of segment in block*/
 	uint16_t numSegmentsAllocated; // number of segments allocated for segs array
 };
